FR-13038-fix-angular-tests-with-16-support
