+++ projects/frontegg-app/src/lib/frontegg-app.module.ts
@@ -2,9 +2,8 @@ import { ModuleWithProviders, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { FronteggAppOptions } from '@frontegg/types';
-import { FronteggAppOptionsClass, FronteggAppService } from './frontegg-app.service';
+import { FronteggAppOptionsClass, FronteggAppService, FronteggLoadGuard } from './frontegg-app.service';
 import { FronteggComponent } from './frontegg.component';
-import { FronteggLoadGuard } from './guards/frontegg-load.guard';
 import { FronteggAuthGuard } from './guards/frontegg-auth.guard';
 import { FronteggAuthService } from './frontegg-auth.service';
 import { AuthorizedContentDirective } from './directives/authorized-content.directive';
