+++ projects/frontegg-app/src/public-api.ts
@@ -10,5 +10,4 @@ export { FronteggEntitlementsService } from './lib/frontegg-entitlements.service
 export { FronteggSubscriptionService } from './lib/frontegg-subscription.services';
 export * from './lib/directives/authorized-content.directive';
 export * from './lib/guards/frontegg-auth.guard';
-export * from './lib/guards/frontegg-load.guard';
 export { ContextHolder } from '@frontegg/rest-api';
