+++ projects/frontegg-app/tsconfig.lib.prod.json
@@ -5,7 +5,6 @@
     "declarationMap": false
   },
   "angularCompilerOptions": {
-    "enableIvy": false,
     "compilationMode": "partial"
   }
 }
